+++ lib/src/http_client_io.dart
@@ -51,8 +51,8 @@ class HttpClientRequesterIO extends HttpClientRequester {
 
     if (responseHeaderWithToken != null) {
       var accessToken = resp.getResponseHeader(responseHeaderWithToken);
-      if (accessToken != null) {
-        client.authorization = BearerCredential(accessToken);
+      if (accessToken != null && accessToken.isNotEmpty) {
+        client.authorization = Authorization.fromCredential( BearerCredential(accessToken) ) ;
       }
     }